requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==3.0.0
boto3==1.35.73
botocore==1.35.73
certifi==2024.8.30
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jiter==0.8.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numexpr==2.10.2
numpy==2.2.1
openai==1.56.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
s3transfer==0.10.4
six==1.16.0
slack_sdk==3.34.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
watchtower==3.3.1
wcwidth==0.2.13
PyMuPDF==1.23.8
Pillow==10.3.0
langfuse==2.57.12
google-generativeai==0.8.4
google-api-python-client==2.160.0
instructor==1.7.2
jsonref==1.1.0